Add files via upload
Case - H_Accounting 
Hult International Business School:

Objectives:  Students will be given access (read-only) to an existing accounting database H_Accounting. Students will need to build queries and insert results into temporary tables through a stored procedure to create the Profit and Loss (P&L) and Balance Sheet (B/S) statements. 

The stored procedure must achieve the following:

Receive as an argument the year for which the user wants to produce the P&L and B/S statements
Do the necessary calculations to obtain the net profit or loss for that year
Use that value, to produce the B/S
Demonstrate that A = L + E and print both the P&L and B/S
Show the % change vs. the previous year for every major line item on the P&L and B/S
Give some headings to each major account line item. And to each section of the P&L and B/S
Add any additional financial metrics and ratios that could be useful to analyze this start-up 
The stored procedure must have comprehensive comments explaining the purpose of each block of code